src/app/components/input/input.component.ts
import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';

@Component({
  selector: 'app-input',
  templateUrl: './input.component.html',
  styleUrls: ['./input.component.scss']
})
export class InputComponent implements OnInit {

  @Input() inputValue: string = '';
  @Input() inputLabel: string = '';
  @Output() inputTextEmitter: EventEmitter<any> = new EventEmitter();

  constructor() { }

  ngOnInit(): void {
  }

  emitText(){
    this.inputTextEmitter.emit(this.inputValue)
  }

}
